(GH-646) Fix vagrant win 1vcpu, port 3389, puppet-agent.log
Remove agent install log only if the log exists.  Avoids this error:

    ==> win2012r2-client: Log will be written to C:/vagrant/puppet-agent.log
    ==> win2012r2-client: Remove-Item : Cannot find path 'C:\vagrant\puppet-agent.log' because it does
    ==> win2012r2-client: not exist.
    ==> win2012r2-client: At C:\tmp\vagrant-shell.ps1:9 char:3
    ==> win2012r2-client: +   Remove-Item $log
    ==> win2012r2-client: +   ~~~~~~~~~~~~~~~~
    ==> win2012r2-client:     + CategoryInfo          : ObjectNotFound: (C:\vagrant\puppet-agent.log:Str
    ==> win2012r2-client:    ing) [Remove-Item], ItemNotFoundException
    ==> win2012r2-client:     + FullyQualifiedErrorId : PathNotFound,Microsoft.PowerShell.Commands.Remov
    ==> win2012r2-client:    eItemCommand

+++ tests/provision_basic_win.ps1
@@ -6,7 +6,7 @@ if ( Get-Command "puppet" -ErrorAction SilentlyContinue ) {
 } else {
   Write-Output "Installing Puppet from $agent_url"
   Write-Output "Log will be written to $log"
-  Remove-Item $log
+  if ( Test-Path $log ) { Remove-Item $log }
   # Install puppet
   Start-Process msiexec.exe -Wait -NoNewWindow -ArgumentList @("/i", "$agent_url", "/qn", "/l*", "$log")
 }
